[fritzboxtr064] Use long instead of int when parsing numbers (#5221)
this fixes an integer overflow for some values returned by the FritzBox

+++ bundles/binding/org.openhab.binding.fritzboxtr064/src/main/java/org/openhab/binding/fritzboxtr064/internal/FritzboxTr064Binding.java
@@ -217,9 +217,9 @@ public class FritzboxTr064Binding extends AbstractActiveBinding<FritzboxTr064Bin
                 eventPublisher.postUpdate(itemName, newState);
             } else if (itemType.isAssignableFrom(NumberItem.class)) { // number items e.g. TAM messages
                 // tr064 retrieves only Strings, trying to parse value returned
-                int val = 0;
+                long val = 0;
                 try {
-                    val = Integer.parseInt(tr064result);
+                    val = Long.parseLong(tr064result);
                 } catch (NumberFormatException ex) {
                     val = -1; // indicate error as -1
                 }
